Rename forward/point_mass.py to forward/point.py
