+++ harmonica/tests/test_point_mass.py
@@ -13,7 +13,7 @@ import pytest
 import verde as vd
 
 from ..constants import GRAVITATIONAL_CONST
-from ..forward.point_mass import point_mass_gravity
+from ..forward.point import point_mass_gravity
 from ..forward.utils import distance_cartesian
 
 from .utils import run_only_with_numba
